Fix upgrade script

+++ plugins/tracker/db/mysql/updates/2012/201206221459_add_postaction_field_float_table.php
@@ -33,7 +33,7 @@ class b201206221459_add_postaction_field_float_table extends ForgeUpgrade_Bucket
                     id            INT(11) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
                     transition_id INT(11) NOT NULL,
                     field_id      INT(11) UNSIGNED DEFAULT NULL,
-                    value         INT(11) DEFAULT NULL,
+                    value         FLOAT(10,4) DEFAULT NULL,
                     
                     INDEX idx_wf_transition_id (transition_id)
                 );";
